pyproject.toml
[project]
name = "oracle"
version = "0.1.0"
description = "Time series forecasting evaluation framework"
requires-python = ">=3.12,<3.13"
dependencies = [
    "chronos-forecasting",
    "pandas",
    "numpy",
    "matplotlib",
    "OMIEData",
    "pmdarima",
    "torch",
    "scikit-learn",
    "statsmodels",
    "scipy",
    "tqdm",
    "gluonts",
    "bs4>=0.0.2",
    "meteostat>=2.1.4",
    "dotenv>=0.9.9",
    "entsoe-py>=0.7.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["wrappers"]
